+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Fixed an issue if script is not calld as root
     * Previous to this fix, all arguments where lost (like `-f`)
+* Centralized code by creating `_create_directory_of_exit` an `_remove_path_or_exit`
 
 ## [2.3.0](https://github.com/stevleibelt/arch-linux-live-cd-iso-with-zfs/tree/2.3.0) - released at 20220328